+++ src/platform.ts
@@ -133,6 +133,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
 
       // Curtain Config Options
       if (this.config.options?.curtain) {
+        !this.config.options.curtain.disable_group;
         this.config.options.curtain.set_minStep;
         this.config.options.curtain.set_min;
         this.config.options.curtain.set_max;
@@ -203,7 +204,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
             if (this.config.devicediscovery) {
               this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
             }
-            this.createHumidifier(device, devices);
+            this.createHumidifier(device);
             break;
           case 'Hub Mini':
             if (this.config.devicediscovery) {
@@ -219,19 +220,19 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
             if (this.config.devicediscovery) {
               this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
             }
-            this.createBot(device, devices);
+            this.createBot(device);
             break;
           case 'Meter':
             if (this.config.devicediscovery) {
               this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
             }
-            this.createMeter(device, devices);
+            this.createMeter(device);
             break;
           case 'Curtain':
             if (this.config.devicediscovery) {
               this.log.info('Discovered %s %s', device.deviceName, device.deviceType);
             }
-            this.createCurtain(device, devices);
+            this.createCurtain(device);
             break;
           case 'Remote':
             if (this.config.devicediscovery) {
@@ -336,7 +337,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     }
   }
 
-  private async createHumidifier(device: device, devices: deviceResponses) {
+  private async createHumidifier(device: device) {
     const uuid = this.api.hap.uuid.generate(`${device.deviceName}-${device.deviceId}-${device.deviceType}`);
 
     // see if an accessory with the same uuid has already been registered and restored from
@@ -345,7 +346,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
+      if (!this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService) {
         this.log.info(
           'Restoring existing accessory from cache: %s DeviceID: %s',
           existingAccessory.displayName,
@@ -362,7 +363,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService) {
       // the accessory does not yet exist, so we need to create it
       this.log.info('Adding new accessory: %s %s DeviceID: %s', device.deviceName, device.deviceType, device.deviceId);
 
@@ -383,7 +384,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -394,7 +395,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     }
   }
 
-  private async createBot(device: device, devices: deviceResponses) {
+  private async createBot(device: device) {
     const uuid = this.api.hap.uuid.generate(`${device.deviceName}-${device.deviceId}-${device.deviceType}`);
 
     // see if an accessory with the same uuid has already been registered and restored from
@@ -403,7 +404,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
+      if (!this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService) {
         this.log.info(
           'Restoring existing accessory from cache: %s DeviceID: %s',
           existingAccessory.displayName,
@@ -420,7 +421,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService) {
       // the accessory does not yet exist, so we need to create it
       this.log.info('Adding new accessory: %s %s DeviceID: %s', device.deviceName, device.deviceType, device.deviceId);
 
@@ -444,7 +445,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -455,7 +456,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     }
   }
 
-  private async createMeter(device: device, devices: deviceResponses) {
+  private async createMeter(device: device) {
     const uuid = this.api.hap.uuid.generate(`${device.deviceName}-${device.deviceId}-${device.deviceType}`);
 
     // see if an accessory with the same uuid has already been registered and restored from
@@ -464,7 +465,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
+      if (!this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService) {
         this.log.info(
           'Restoring existing accessory from cache: %s DeviceID: %s',
           existingAccessory.displayName,
@@ -481,7 +482,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
+    } else if (!this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService) {
       // the accessory does not yet exist, so we need to create it
       this.log.info('Adding new accessory: %s %s DeviceID: %s', device.deviceName, device.deviceType, device.deviceId);
 
@@ -502,7 +503,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -513,7 +514,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     }
   }
 
-  private async createCurtain(device: device, devices: deviceResponses) {
+  private async createCurtain(device: device) {
     const uuid = this.api.hap.uuid.generate(`${device.deviceName}-${device.deviceId}-${device.deviceType}`);
 
     // see if an accessory with the same uuid has already been registered and restored from
@@ -522,7 +523,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
 
     if (existingAccessory) {
       // the accessory already exists
-      if (!this.config.options?.hide_device.includes(device.deviceId) && devices.statusCode === 100) {
+      if (this.isCurtainGrouped(device)) {
         this.log.info(
           'Restoring existing accessory from cache: %s DeviceID: %s',
           existingAccessory.displayName,
@@ -539,14 +540,21 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       } else {
         this.unregisterPlatformAccessories(existingAccessory);
       }
-    } else if (!this.config.options?.hide_device.includes(device.deviceId)) {
+    } else if (this.isCurtainGrouped(device)) {
       // the accessory does not yet exist, so we need to create it
       this.log.info('Adding new accessory: %s %s DeviceID: %s', device.deviceName, device.deviceType, device.deviceId);
-
-      if (device.group) {
+      if (device.group && !this.config.options?.curtain?.disable_group) {
         this.log.warn(
-          'Your Curtains are Grouped, you must hide your secondary curtian using the Hidden Device Setting',
+          'Your Curtains are grouped, Secondary curtain automatically hidden. Main Curtain: %s, DeviceID: %s',
+          device.deviceName,
+          device.deviceId,
         );
+      } else {
+        if (device.master) {
+          this.log.warn('Main Curtain: %s, DeviceID: %s', device.deviceName, device.deviceId);
+        } else {
+          this.log.warn('Secondary Curtain: %s, DeviceID: %s', device.deviceName, device.deviceId);
+        }
       }
 
       // create a new accessory
@@ -566,7 +574,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -574,6 +582,14 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
           device.deviceId,
         );
       }
+    }
+  }
+
+  private isCurtainGrouped(device: device) {
+    if (device.group && !this.config.options?.curtain?.disable_group) {
+      return device.master && !this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService;
+    } else {
+      return !this.config.options?.hide_device.includes(device.deviceId) && device.enableCloudService;
     }
   }
 
@@ -629,7 +645,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.publishExternalAccessories(PLUGIN_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -687,7 +703,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -745,7 +761,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -803,7 +819,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -861,7 +877,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -919,7 +935,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -977,7 +993,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,
@@ -1035,7 +1051,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [accessory]);
       this.accessories.push(accessory);
     } else {
-      if (!this.config.options?.hide_device.includes(device.deviceId)) {
+      if (this.config.devicediscovery) {
         this.log.error(
           'Unable to Register new device: %s %s - %s',
           device.deviceName,